requirements.txt
pika==0.9.13
pymongo==2.6.3
python-dateutil==2.2